Create Task 3 weather report
The project makes use of various data sources to gather weather information such as temperature, precipitation, pressure, and other important meteorological data. This data is then processed and analyzed using various data analysis and visualization tools such as SQL, Excel, Power BI/ Tableau and Python. The results of this analysis are then presented to the users through interactive  graphs, making it easier for them to understand and analyze weather trends and patterns.